refactor: SignUpStageThree
